requirements.txt
Django==2.2.13
django-cleanup==4.0.0
django-ckeditor==6.0.0
translate==3.5.0
pytz==2018.4
six==1.11.0
requests==2.7.0
unicodecsv==0.14.1
